requirements.txt
openai
PyGithub
python-dotenv
colorama